move new hash and serialization code to v2
