+++ api/hashing/v2/serialize/store_chunk.go
@@ -1,4 +1,4 @@
-package serialization
+package serialize
 
 import (
 	"encoding/binary"
