+++ api/hashing/v2/serialize/store_chunk_compat_test.go
@@ -1,4 +1,4 @@
-package serialization
+package serialize
 
 import (
 	"bytes"
