+++ api/hashing/v2/serialize/store_chunk_v2.go
@@ -1,4 +1,4 @@
-package serialization
+package serialize
 
 import (
 	"bytes"
