+++ api/hashing/v2/store_chunk_request.go
@@ -4,12 +4,12 @@ import (
 	"fmt"
 
 	grpc "github.com/Layr-Labs/eigenda/api/grpc/validator"
-	"github.com/Layr-Labs/eigenda/api/hashing/serialization"
+	"github.com/Layr-Labs/eigenda/api/hashing/v2/serialize"
 	"golang.org/x/crypto/sha3"
 )
 
 func HashStoreChunksRequest_Canonical(request *grpc.StoreChunksRequest) ([]byte, error) {
-	canonicalRequest, err := serialization.SerializeStoreChunksRequest(request)
+	canonicalRequest, err := serialize.SerializeStoreChunksRequest(request)
 	if err != nil {
 		return nil, fmt.Errorf("failed to serialize store chunks request: %w", err)
 	}
@@ -19,7 +19,7 @@ func HashStoreChunksRequest_Canonical(request *grpc.StoreChunksRequest) ([]byte,
 }
 
 func HashStoreChunksRequest_V2_Canonical(request *grpc.StoreChunksRequest) ([]byte, error) {
-	canonicalRequest, err := serialization.SerializeStoreChunksRequestV2(request)
+	canonicalRequest, err := serialize.SerializeStoreChunksRequestV2(request)
 	if err != nil {
 		return nil, fmt.Errorf("failed to serialize store chunks request: %w", err)
 	}
